requirements.txt
dnspython
python-whois
requests
ping3
fastapi
uvicorn
httpx
google-genai
